+++ tests/Fixtures/Annotation/Route/Attribute/GlobalDefaultsClass.php
@@ -1,0 +1,34 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of Flight Routing.
+ *
+ * PHP version 7.1 and above required
+ *
+ * @author    Divine Niiquaye Ibok <divineibok@gmail.com>
+ * @copyright 2019 Biurad Group (https://biurad.com/)
+ * @license   https://opensource.org/licenses/BSD-3-Clause License
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Flight\Routing\Tests\Fixtures\Annotation\Route\Attribute;
+
+use Flight\Routing\Annotation\Route;
+
+#[Route(path: '/defaults', name: 'attribute_']
+class GlobalDefaultsClass
+{
+    #[Route(path: '/specific-name', name: 'specific_name', methods: ['GET'])]
+    public function withName()
+    {
+    }
+
+    #[Route(path: '/specific-none', methods: ['GET', 'HEAD']]
+    public function noName()
+    {
+    }
+}